added note about ftpdump
if you do want to use the new ftpdump script all you have to do is replace the ftpdump file and add two spaces after the

ip=

and add the port you want to use.

the program will make a copy of ftpdump named yourftpdump. it keeps the ftpdump persistent for updates to his script.